lattice_qcl_to_south/requirements-dev.txt
aiohttp~=3.8.4
fastapi==0.100.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.7
fastapi-utils==0.2.1
typing_extensions==4.5.0
typing== 3.7.4.1
python-decouple==3.8
aioredis==2.0.1
motor==3.2.0
pymongo==4.4.1
ipdb==0.13.13